backend/requirements.txt
annotated-types==0.7.0
anyio==3.7.1
bcrypt==4.3.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.104.1
google-auth==2.23.4
google-auth-oauthlib==1.1.0
h11==0.16.0
httpx==0.27.0
idna==3.10
motor==3.3.2
oauthlib==3.3.1
passlib==1.7.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
pymongo==4.6.0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.20
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
scalar_fastapi==1.4.0
six==1.17.0
sniffio==1.3.1
starlette==0.27.0
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.24.0
